it works :)
Items are linked to list entires when the list is saved.
Stores are found when the user navigates to the /results/ page.
The algorithm is by no means perfect, but it seems to get the job done alright, stores are selected first based on number missing, then based on price. This seems to give the most desirable result as of now.